Removed MagmaHeart.Core.Cameras namespaces. All camera related stuff moved to MagmaHeart.Core.Entities.PlayableCharacters
