--- Assets/_Scripts/Core/CameraControls.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7ebf27eb7cb85a942a22a6cb842f239d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
