--- Assets/_Scripts/Core/CameraControls/ActionCameraBehaviour.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 87088f6a8d0388f45bf503b7ea157ff3
