--- Assets/_Scripts/Core/CameraControls/CameraController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3020f295bc28f7d4194df76b5f1b15f8
