--- Assets/_Scripts/Core/CameraControls/ICameraBehaviour.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fc3167174fbaa5044a857642596949ee
