--- Assets/_Scripts/Core/CameraControls/TurnBasedCameraBehaviour.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c91eaab767a59d141abc4a666c238bbb
