+++ Assets/_Scripts/Core/Entities/PlayableCharacters/ActionCameraBehaviour.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d313bfcc3f017d547a66de03f4f875f5
