+++ Assets/_Scripts/Core/Entities/PlayableCharacters/ICameraBehaviour.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 544543004b82a454781eb1ed7103e21f
